refactor(shielded): extract ShieldedHeaderVerifier mirroring existing verifier headers
